Split visitors into their own files, have collector skip any variables that have already been resolved in some way
